4/11/17
Separated parts of the UI Manager.
Each section is broken up into UIWorkShop/UICrafting/UIActions.
